LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/moving-between-documents/move-back-iframe-parse-error-external-classic-expected.txt

CONSOLE MESSAGE: SyntaxError: Left side of assignment is not a reference.


PASS Sanity check around top-level Window
PASS Eval: Move parse-error external classic script to iframe move-back
FAIL <script> load: Move parse-error external classic script to iframe move-back assert_unreached: Script load event fired unexpectedly Reached unreachable code
PASS <script> error: Move parse-error external classic script to iframe move-back
FAIL window error: Move parse-error external classic script to iframe move-back assert_unreached: Window error event shouldn't fired on source window Reached unreachable code
